pyproject.toml
[project]
name = "imaging-server-kit"
dynamic = ["version"]
description = "Run image processing algorithms in a FastAPI server."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Mallory Wittwer", email = "mallory.wittwer@epfl.ch" }]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "tifffile",
    "numpy",
    "scikit-image",
    "geojson",
    "docker",  # For the autostartup regsitry
    "jinja2",  # For the /info route
    "pyyaml",  # For the /info route
]

[project.urls]
homepage = "https://gitlab.com/epfl-center-for-imaging/imaging-server-kit"
repository = "https://gitlab.com/epfl-center-for-imaging/imaging-server-kit"

[build-system]
requires = ["setuptools>=42.0.0", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.setuptools_scm]
write_to = "src/imaging_server_kit/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
